requirements.txt
asgiref==3.8.1
asttokens==3.0.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
decorator==5.1.1
Django==5.1.5
django-formtools==2.5.1
emoji==2.14.1
exceptiongroup==1.2.2
executing==2.2.0
idna==3.10
ipython==8.32.0
jedi==0.19.2
matplotlib-inline==0.1.7
numpy==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
prompt_toolkit==3.0.50
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.19.1
pyOpenSSL==25.0.0
pytz==2025.1
requests==2.32.3
sqlparse==0.5.3
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
urllib3==2.3.0
wcwidth==0.2.13
gunicorn==19.7.1
whitenoise
dj_database_url
psycopg2-binary
django-environ
dj-database-url